refactor(map_loader): modularization
